refactor v2 collection predicate operators
